introduce an optional ignoreFn option
and change the honorIndex boolean arg to an options object

This allows people to selectably ignore certain files. The problem
caused my build to fail as I had a dependency on a npm package which had
a file called the same as the package, and this plugin took precedence
over resolving through the main property in package.json.

This allows me to short circuit the plugin for certain requests.